S302T700-525 #time 30m Return 404 when deleting profile pic of user with no profile pic ?fix ?test ?document - Also adds the endpoint to the API spec

+++ backend/src/main/java/com/seng302/wasteless/controller/ImageController.java
@@ -307,6 +307,10 @@ public class ImageController {
     public ResponseEntity<Object> deleteUserImage(@PathVariable("userId") Integer userId) {
         User userForImage = getUserToModify(userId);
 
+        if (userForImage.getProfileImage() == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "The given user does not have a profile image");
+        }
+
         userService.deleteUserImage(userForImage);
 
         return ResponseEntity.status(HttpStatus.OK).build();
